ai-service/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2

# AI / ML
ultralytics==8.3.2
opencv-python-headless==4.10.0.84
easyocr==1.7.2
numpy==1.26.4
Pillow==10.4.0
torch==2.4.1
torchvision==0.19.1

# Async & Performance
aiofiles==24.1.0
asyncio==3.4.3
aiohttp==3.10.5

# Logging & Monitoring
loguru==0.7.2
prometheus-fastapi-instrumentator==7.0.0

# Utils
python-dotenv==1.0.1
tenacity==9.0.0
